BLE: Move security DB into PAL.
